[Next Gen] refactor(codegen): split JS-string escaping out of helpers into escape
